requirements.txt
gradio
chromadb
openai
pandas
python-dotenv